pyproject.toml
[project]
name = "mcu-rag"
version = "0.1.0"
description = "MarvelMind — MCU RAG Showcase App for AI Course"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40",
    "chromadb>=0.5",
    "llama-index-core>=0.12",
    "llama-index-vector-stores-chroma>=0.4",
    "llama-index-embeddings-ollama>=0.5",
    "llama-index-llms-ollama>=0.5",
    "llama-index-readers-file>=0.4",
    "ollama>=0.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0",
    "datasets>=2.20",
    "Wikipedia-API>=0.7",
    "beautifulsoup4>=4.12",
    "requests>=2.32",
    "tqdm>=4.66",
    "numpy>=1.26",
    "pandas>=2.2",
    "plotly>=5.20",
    "httpx>=0.27",
    # Docling: optional but recommended for high-quality PDF parsing
    # Uncomment to install: uv add docling
    # "docling>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcu_rag"]

[tool.uv]
dev-dependencies = []
